backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1

# Redis & Queue
redis>=5.0.1
celery>=5.3.6

# Shopify Integration
ShopifyAPI>=12.5.0

# AI/LLM
anthropic>=0.18.1

# HTTP Client
httpx>=0.27.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0

# Durable Execution
temporalio>=1.5.0
